prettier

+++ contracts/lib/proto/Pb.sol
@@ -4,11 +4,18 @@ pragma solidity >=0.5.0;
 
 // runtime proto sol library
 library Pb {
-    enum WireType { Varint, Fixed64, LengthDelim, StartGroup, EndGroup, Fixed32 }
+    enum WireType {
+        Varint,
+        Fixed64,
+        LengthDelim,
+        StartGroup,
+        EndGroup,
+        Fixed32
+    }
 
     struct Buffer {
-        uint idx;  // the start index of next read. when idx=b.length, we're done
-        bytes b;   // hold serialized proto msg, readonly
+        uint256 idx; // the start index of next read. when idx=b.length, we're done
+        bytes b; // hold serialized proto msg, readonly
     }
 
     // create a new in-memory Buffer object from raw msg bytes
@@ -23,19 +30,19 @@ library Pb {
     }
 
     // decode current field number and wiretype
-    function decKey(Buffer memory buf) internal pure returns (uint tag, WireType wiretype) {
-        uint v = decVarint(buf);
+    function decKey(Buffer memory buf) internal pure returns (uint256 tag, WireType wiretype) {
+        uint256 v = decVarint(buf);
         tag = v / 8;
         wiretype = WireType(v & 7);
     }
 
     // count tag occurrences, return an array due to no memory map support
-	// have to create array for (maxtag+1) size. cnts[tag] = occurrences
-	// should keep buf.idx unchanged because this is only a count function
-    function cntTags(Buffer memory buf, uint maxtag) internal pure returns (uint[] memory cnts) {
-        uint originalIdx = buf.idx;
-        cnts = new uint[](maxtag+1);  // protobuf's tags are from 1 rather than 0
-        uint tag;
+    // have to create array for (maxtag+1) size. cnts[tag] = occurrences
+    // should keep buf.idx unchanged because this is only a count function
+    function cntTags(Buffer memory buf, uint256 maxtag) internal pure returns (uint256[] memory cnts) {
+        uint256 originalIdx = buf.idx;
+        cnts = new uint256[](maxtag + 1); // protobuf's tags are from 1 rather than 0
+        uint256 tag;
         WireType wire;
         while (hasMore(buf)) {
             (tag, wire) = decKey(buf);
@@ -46,18 +53,18 @@ library Pb {
     }
 
     // read varint from current buf idx, move buf.idx to next read, return the int value
-    function decVarint(Buffer memory buf) internal pure returns (uint v) {
-        bytes10 tmp;  // proto int is at most 10 bytes (7 bits can be used per byte)
-        bytes memory bb = buf.b;  // get buf.b mem addr to use in assembly
-        v = buf.idx;  // use v to save one additional uint variable
+    function decVarint(Buffer memory buf) internal pure returns (uint256 v) {
+        bytes10 tmp; // proto int is at most 10 bytes (7 bits can be used per byte)
+        bytes memory bb = buf.b; // get buf.b mem addr to use in assembly
+        v = buf.idx; // use v to save one additional uint variable
         assembly {
             tmp := mload(add(add(bb, 32), v)) // load 10 bytes from buf.b[buf.idx] to tmp
         }
-        uint b; // store current byte content
+        uint256 b; // store current byte content
         v = 0; // reset to 0 for return value
-        for (uint i=0; i<10; i++) {
+        for (uint256 i = 0; i < 10; i++) {
             assembly {
-                b := byte(i, tmp)  // don't use tmp[i] because it does bound check and costs extra
+                b := byte(i, tmp) // don't use tmp[i] because it does bound check and costs extra
             }
             v |= (b & 0x7F) << (i * 7);
             if (b & 0x80 == 0) {
@@ -70,19 +77,19 @@ library Pb {
 
     // read length delimited field and return bytes
     function decBytes(Buffer memory buf) internal pure returns (bytes memory b) {
-        uint len = decVarint(buf);
-        uint end = buf.idx + len;
-        require(end <= buf.b.length);  // avoid overflow
+        uint256 len = decVarint(buf);
+        uint256 end = buf.idx + len;
+        require(end <= buf.b.length); // avoid overflow
         b = new bytes(len);
-        bytes memory bufB = buf.b;  // get buf.b mem addr to use in assembly
-        uint bStart;
-        uint bufBStart = buf.idx;
+        bytes memory bufB = buf.b; // get buf.b mem addr to use in assembly
+        uint256 bStart;
+        uint256 bufBStart = buf.idx;
         assembly {
             bStart := add(b, 32)
             bufBStart := add(add(bufB, 32), bufBStart)
         }
-        for (uint i=0; i<len; i+=32) {
-            assembly{
+        for (uint256 i = 0; i < len; i += 32) {
+            assembly {
                 mstore(add(bStart, i), mload(add(bufBStart, i)))
             }
         }
@@ -90,20 +97,20 @@ library Pb {
     }
 
     // return packed ints
-    function decPacked(Buffer memory buf) internal pure returns (uint[] memory t) {
-        uint len = decVarint(buf);
-        uint end = buf.idx + len;
-        require(end <= buf.b.length);  // avoid overflow
+    function decPacked(Buffer memory buf) internal pure returns (uint256[] memory t) {
+        uint256 len = decVarint(buf);
+        uint256 end = buf.idx + len;
+        require(end <= buf.b.length); // avoid overflow
         // array in memory must be init w/ known length
         // so we have to create a tmp array w/ max possible len first
-        uint[] memory tmp = new uint[](len);
-        uint i = 0; // count how many ints are there
+        uint256[] memory tmp = new uint256[](len);
+        uint256 i = 0; // count how many ints are there
         while (buf.idx < end) {
             tmp[i] = decVarint(buf);
             i++;
         }
-        t = new uint[](i); // init t with correct length
-        for (uint j=0; j<i; j++) {
+        t = new uint256[](i); // init t with correct length
+        for (uint256 j = 0; j < i; j++) {
             t[j] = tmp[j];
         }
         return t;
@@ -111,23 +118,28 @@ library Pb {
 
     // move idx pass current value field, to beginning of next tag or msg end
     function skipValue(Buffer memory buf, WireType wire) internal pure {
-        if (wire == WireType.Varint) { decVarint(buf); }
-        else if (wire == WireType.LengthDelim) {
-            uint len = decVarint(buf);
+        if (wire == WireType.Varint) {
+            decVarint(buf);
+        } else if (wire == WireType.LengthDelim) {
+            uint256 len = decVarint(buf);
             buf.idx += len; // skip len bytes value data
-            require(buf.idx <= buf.b.length);  // avoid overflow
-        } else { revert(); }  // unsupported wiretype
+            require(buf.idx <= buf.b.length); // avoid overflow
+        } else {
+            revert();
+        } // unsupported wiretype
     }
 
     // type conversion help utils
-    function _bool(uint x) internal pure returns (bool v) {
+    function _bool(uint256 x) internal pure returns (bool v) {
         return x != 0;
     }
 
     function _uint256(bytes memory b) internal pure returns (uint256 v) {
-        require(b.length <= 32);  // b's length must be smaller than or equal to 32
-        assembly { v := mload(add(b, 32)) }  // load all 32bytes to v
-        v = v >> (8 * (32 - b.length));  // only first b.length is valid
+        require(b.length <= 32); // b's length must be smaller than or equal to 32
+        assembly {
+            v := mload(add(b, 32))
+        } // load all 32bytes to v
+        v = v >> (8 * (32 - b.length)); // only first b.length is valid
     }
 
     function _address(bytes memory b) internal pure returns (address v) {
@@ -137,32 +149,44 @@ library Pb {
     function _addressPayable(bytes memory b) internal pure returns (address payable v) {
         require(b.length == 20);
         //load 32bytes then shift right 12 bytes
-        assembly { v := div(mload(add(b, 32)), 0x1000000000000000000000000) }
+        assembly {
+            v := div(mload(add(b, 32)), 0x1000000000000000000000000)
+        }
     }
 
     function _bytes32(bytes memory b) internal pure returns (bytes32 v) {
         require(b.length == 32);
-        assembly { v := mload(add(b, 32)) }
+        assembly {
+            v := mload(add(b, 32))
+        }
     }
 
     // uint[] to uint8[]
-    function uint8s(uint[] memory arr) internal pure returns (uint8[] memory t) {
+    function uint8s(uint256[] memory arr) internal pure returns (uint8[] memory t) {
         t = new uint8[](arr.length);
-        for (uint i = 0; i < t.length; i++) { t[i] = uint8(arr[i]); }
+        for (uint256 i = 0; i < t.length; i++) {
+            t[i] = uint8(arr[i]);
+        }
     }
 
-    function uint32s(uint[] memory arr) internal pure returns (uint32[] memory t) {
+    function uint32s(uint256[] memory arr) internal pure returns (uint32[] memory t) {
         t = new uint32[](arr.length);
-        for (uint i = 0; i < t.length; i++) { t[i] = uint32(arr[i]); }
+        for (uint256 i = 0; i < t.length; i++) {
+            t[i] = uint32(arr[i]);
+        }
     }
 
-    function uint64s(uint[] memory arr) internal pure returns (uint64[] memory t) {
+    function uint64s(uint256[] memory arr) internal pure returns (uint64[] memory t) {
         t = new uint64[](arr.length);
-        for (uint i = 0; i < t.length; i++) { t[i] = uint64(arr[i]); }
+        for (uint256 i = 0; i < t.length; i++) {
+            t[i] = uint64(arr[i]);
+        }
     }
 
-    function bools(uint[] memory arr) internal pure returns (bool[] memory t) {
+    function bools(uint256[] memory arr) internal pure returns (bool[] memory t) {
         t = new bool[](arr.length);
-        for (uint i = 0; i < t.length; i++) { t[i] = arr[i]!=0; }
+        for (uint256 i = 0; i < t.length; i++) {
+            t[i] = arr[i] != 0;
+        }
     }
 }
